refactor(select): StyledReactSelect to private
- Update StyledReactSelect to private component
- Update import path
